requirements.txt
djangorestframework~=3.12.4
requests~=2.26.0
django-filter

Django~=3.2.9